some changes

+++ src/main/java/com/veterinaryclinic/Services/PatientsService.java
@@ -1,10 +1,11 @@
 package com.veterinaryclinic.Services;
 
-import java.util.ArrayList;
+import java.util.List;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
-import com.veterinaryclinic.Controllers.AppointmentController;
-import com.veterinaryclinic.Controllers.PatientsController;
+
+import com.veterinaryclinic.Models.Patient;
 import com.veterinaryclinic.Repositories.IPatientsRepository;
 
 @Service
@@ -13,28 +14,36 @@ public class PatientsService {
     @Autowired
     IPatientsRepository patientsRepository;
 
-    public ArrayList<PatientsController> getAllPatients() {
-        return patientsRepository.getAllPatients();
+    public List<Patient> getAllPatients() {
+        return patientsRepository.findAll();
     }
 
-    public ArrayList<PatientsController> getPatientsById(Long id) {
-        return patientsRepository.getPatientsById();
+    public Patient getPatientById(Long id) {
+        return patientsRepository.findById(id).orElse(null);
     }
 
-    public PatientsController savePatients(PatientsController patients) {
-        return patientsRepository.savePatients();
+    public Patient savePatient(Patient patient) {
+        return patientsRepository.save(patient);
     }
 
-    public void updatePatients(PatientsController patients) {
-        return patientsRepository.updatePatients();
+    public Patient updatePatient(Long id, Patient updatedPatient) {
+        if (patientsRepository.existsById(id)) {
+            updatedPatient.setId(id);
+            return patientsRepository.save(updatedPatient);
+        }
+        return null;
     }
 
-    public String deletePatients(Long id) {
-        try{
-            patientsRepository.deleteById(id);
-            return "todo ok";
-        } catch (Exception error){
-        return "todo mal"; }
+    public String deletePatient(Long id) {
+        try {
+            if (patientsRepository.existsById(id)) {
+                patientsRepository.deleteById(id);
+                return "all ok";
+            } else {
+                return "patient not found";
+            }
+        } catch (Exception e) {
+            return "error: " + e.getMessage();
+        }
     }
-
-    }
+}
